Cast enum valuez to int

+++ WinPath.Tests/EventType.cs
@@ -14,11 +14,11 @@ namespace WinPath.Tests
 		public void AllValuesAreInOrder()
 		{
 			bool valuesAreInOrder =
-				(HandleEventType.NoValue == 0)
-			  + (HandleEventType.UserPath == 1)
-			  + (HandleEventType.SystemPath == 2)
-			  + (HandleEventType.UserAndSystemPath == 3)
-			  + (HandleEventType.NoUserOrSystemPath == 4);
+				((int)HandleEventType.NoValue == 0)
+			  + ((int)HandleEventType.UserPath == 1)
+			  + ((int)HandleEventType.SystemPath == 2)
+			  + ((int)HandleEventType.UserAndSystemPath == 3)
+			  + ((int)HandleEventType.NoUserOrSystemPath == 4);
 			Assert.True(valuesAreInOrder);
 		}
 	}
